Rating displays till one decimal point
ISSUE ID:ACUMOS-2612
Description:IST: Rating is not displaying till one decimal point

+++ acumos-portal-be/src/main/java/org/acumos/portal/be/util/PortalUtils.java
@@ -159,7 +159,7 @@ public class PortalUtils {
 			mlSolution.setViewCount(mlpSolution.getViewCount().intValue());
 		}
 		if(mlpSolution.getRatingAverageTenths() != null) {
-			mlSolution.setSolutionRatingAvg(mlpSolution.getRatingAverageTenths() / 10);
+			mlSolution.setSolutionRatingAvg(mlpSolution.getRatingAverageTenths().floatValue() / 10);
 		}
 		mlSolution.setLastDownload(mlpSolution.getLastDownload());
 		if (mlpSolution.isFeatured() != null) {
